feat: make reachability human friendly
At the moment, the output shows reachability as function or package or with no info. This change makes the wording conistent with the GUI: reachable/ potentially-reachable / no-info
